+++ data_adapter/databus.py
@@ -2,6 +2,7 @@ import json
 import logging
 import os
 import pathlib
+import re
 import urllib.parse
 from typing import List, Union
 
@@ -32,7 +33,8 @@ def download_artifact(artifact_file: str, filename: Union[pathlib.Path, str]):
     if response.status_code != 200:
         raise FileNotFoundError(f"Could not find artifact file '{artifact_file}'")
     with open(os.path.join(filename), "wb") as f:
-        f.write(response.content)
+        # Replace single quotes with two double quotes to enable arrays and JSON/dicts in CSV cells:
+        f.write(re.sub(b"'", b'""', response.content))
 
 
 def get_artifact_filenames(artifact: str, version: str) -> List[str]:
@@ -141,11 +143,16 @@ def download_collection(collection_url: str, force_download=False):
         URL of collection on databus
     force_download : bool
         Downloads the latest versions, even if version is already present
+
+    Raises
+    ------
+    CollectionError
+        If version of collection metadata differs
     """
     logging.info(f"Downloading collection from URL '{collection_url}'...")
     collection_name = collection_url.rstrip("/").split("/")[-1]
     collection_dir = settings.COLLECTIONS_DIR / collection_name
-    collection_meta = {}
+    collection_meta = {"version": settings.COLLECTION_META_VERSION, "artifacts": {}}
     if collection_dir.exists():
         if (collection_dir / settings.COLLECTION_JSON).exists():
             with open(
@@ -154,6 +161,12 @@ def download_collection(collection_url: str, force_download=False):
                 collection_meta = json.load(collection_json_file)
     else:
         collection_dir.mkdir()
+
+    if collection_meta.get("version", None) != settings.COLLECTION_META_VERSION:
+        raise data_adapter.collection.CollectionError(
+            "Collection metadata has changed. Please remove collection and re-download. "
+            "Otherwise, strange behaviours could occur."
+        )
 
     artifacts = get_artifacts_from_collection(collection_url)
     artifact_versions = {
@@ -193,13 +206,13 @@ def __download_artifacts(
     """
     for artifact, version in artifact_versions.items():
         group_name = artifact.split("/")[-2]
-        if group_name not in collection_meta:
-            collection_meta[group_name] = {}
+        if group_name not in collection_meta["artifacts"]:
+            collection_meta["artifacts"][group_name] = {}
         artifact_name = artifact.split("/")[-1]
-        if artifact_name not in collection_meta[group_name]:
-            collection_meta[group_name][artifact_name] = {}
+        if artifact_name not in collection_meta["artifacts"][group_name]:
+            collection_meta["artifacts"][group_name][artifact_name] = {}
 
-        latest_version = collection_meta[group_name][artifact_name].get(
+        latest_version = collection_meta["artifacts"][group_name][artifact_name].get(
             "latest_version"
         )
         if not force_download and latest_version and latest_version == version:
@@ -207,7 +220,9 @@ def __download_artifacts(
                 f"Skipping download of {artifact_name=} {version=} as latest version is already present."
             )
             continue
-        collection_meta[group_name][artifact_name]["latest_version"] = version
+        collection_meta["artifacts"][group_name][artifact_name][
+            "latest_version"
+        ] = version
 
         group_dir = collection_dir / group_name
         if not group_dir.exists():
@@ -227,10 +242,10 @@ def __download_artifacts(
             download_artifact(artifact_filename, version_dir / filename)
             if suffix == "json":
                 metadata = core.get_metadata(version_dir / filename)
-                collection_meta[group_name][artifact_name][
+                collection_meta["artifacts"][group_name][artifact_name][
                     "subject"
                 ] = ontology.get_subject(metadata)
-                collection_meta[group_name][artifact_name][
+                collection_meta["artifacts"][group_name][artifact_name][
                     "datatype"
                 ] = data_adapter.collection.get_data_type(metadata)
         logging.info(f"Downloaded {artifact_name=} {version=}.")